+++ src/Beans.xml
@@ -7,12 +7,17 @@
     </bean>
     <bean id="shoppingCartBean" class="ShoppingCartBean">
     </bean>
+
+    <bean id="ReviewBoardBean" class="ReviewBoardBean">
+    </bean>
+
     <bean id="uiBean" class="UIBean">
         <property name="storeBean" ref="storeBean"/>
     </bean>
     <bean id="storeBean" class="StoreBean">
         <constructor-arg index="0" ref="inventoryBean"/>
         <constructor-arg index="1" ref="shoppingCartBean"/>
-        <constructor-arg index="2" ref="uiBean"/>
+        <constructor-arg index="2" ref="ReviewBoardBean"/>
+        <constructor-arg index="3" ref="uiBean"/>
     </bean>
 </beans>
