+++ src/StoreBean.java
@@ -4,12 +4,14 @@ import java.util.Map;
 public class StoreBean {
     InventoryBean inventoryBean;
     ShoppingCartBean shoppingCartBean;
+    ReviewBoardBean reviewBoardBean;
     UIBean uiBean;
 
-    public StoreBean(InventoryBean inventoryBean, ShoppingCartBean shoppingCartBean, UIBean uiBean) {
+    public StoreBean(InventoryBean inventoryBean, ShoppingCartBean shoppingCartBean, ReviewBoardBean reviewBoardBean, UIBean uiBean) {
         this.inventoryBean = inventoryBean;
         this.shoppingCartBean = shoppingCartBean;
         this.uiBean = uiBean;
+        this.reviewBoardBean = reviewBoardBean;
         this.init();
     }
 
@@ -34,6 +36,10 @@ public class StoreBean {
         return this.inventoryBean.getAllProductsStock();
     }
 
+    public Map<String, String> getAllReviews() {
+        return this.reviewBoardBean.getReviews();
+    }
+
     public void addItemToCart(Product product, int quantity) {
         if (product != null) {
             if (this.inventoryBean.getStock(product) >= quantity) {
@@ -52,6 +58,21 @@ public class StoreBean {
             if (quantity != 0) {
                 this.inventoryBean.increaseStock(product, quantity);
                 this.shoppingCartBean.removeItem(product, quantity);
+            }
+        }
+    }
+
+    public void addItemsToReview(String product) {
+        if(product != null) {
+            this.reviewBoardBean.addReview(product);
+        }
+        System.out.println(reviewBoardBean);
+    }
+
+    public void reviewItem(String product, String reviewMessage) {
+        if(product != null){
+            if(reviewMessage != null){
+                this.reviewBoardBean.setReview(product, reviewMessage);
             }
         }
     }
